Fixed minor bugs and refactoring

+++ src/components/DeleteUserButton.js
@@ -1,8 +1,10 @@
 const DeleteUserButton = ({ index, deleteUser }) => {
   const delUser = () => {
-    deleteUser(index);
+    const warning = window.confirm("Are you sure you want to delete this user ?");
+    if (warning) {
+      deleteUser(index);
+    }
   };
-
   return (
     <button
       onClick={(e) => {
